fix(metrics): the segment-info name buffer is sized by a 16-bit field the bounds convention can see (#94)
Scenario: c3621d2 (item B) read the .mts name length into an `int` and rented by it when it is
over 256 bytes. The value is a 16-bit field, so the rent was never larger than 64 KB, but
FileBoundsConventionTests cannot see that through an `int` and FAILED on it in the full Debug
Storage run ("MetricReader.cs:62 sized by 'nameLen'"), which the targeted runs of item B did not
include.

Change: the length is declared `ushort` - the type is the bound, which is how the convention
recognises a narrow size (`ushort x = ...`). No behaviour change: the same value, the same rent.

Test: FileBoundsConventionTests green again (red on c3621d2); MetricCatalogSeedTests and
MetricFormatV3Tests green (13/13 with it, Debug).

+++ src/Ameto.Metrics/Storage/MetricReader.cs
@@ -58,7 +58,7 @@ internal static class MetricReader
         Span<byte> nameHead = stackalloc byte[6];
         if (nameIdxOffset < 0) throw new IOException($"Name index offset {nameIdxOffset} is before the start of {filePath}");
         if (ReadAt(handle, nameHead, nameIdxOffset) < 6) throw new EndOfStreamException();
-        int     nameLen = BinaryPrimitives.ReadUInt16LittleEndian(nameHead[4..]);
+        ushort  nameLen = BinaryPrimitives.ReadUInt16LittleEndian(nameHead[4..]);   // 16 bits: the type is the bound
         byte[]? rented  = nameLen > 256 ? ArrayPool<byte>.Shared.Rent(nameLen) : null;
         try
         {
